% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot the spot of crashes}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{string or integer specifying the state number
corresponing to an existing state in US}

\item{year}{integer specifying year of the dataset ready to be
read in}
}
\value{
Plot of crashes on the map by latitude and longitude
}
\description{
Plot the spot of fatal injuries suffered in motor vehicle
traffic crashes on the map of the state specified by state
number.
}
\details{
\itemize{
 \item{}{Valid years are exclusively 2013, 2014 and 2015, an error
 message will be post and program will be stopped if invalid years
 are inputted}
 \item{}{User is required to install prerequisted package \code{\link{dplyr}}
 and \code{\link{maps}}, an error message will be post and program will be stopped}
 }
}
\examples{
\dontrun{year <- 2013
state.num <- "31"
fars_map_state(state.num,year)}

}
