% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read in CSV file.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{name of CSV file}
}
\value{
Tibble of read-in dataset
}
\description{
Read the CSV file containing American public yearly data
regarding fatal injuries suffered in motor vehicle traffic crashes, then
convert the default dataframe format to tibble.
}
\details{
\itemize{
 \item{}{If the \code{filename} does not exist in the current dictionary, R
will be stopped and pose a warning massage}
 \item{}{User is recommended to initially install package \code{\link{readr}}
and \code{\link{dplyr}}}
}
}
\examples{
\dontrun{fars_read("accident_2013.csv")}

}
