% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{year}{input from user}
}
\value{
nothing
}
\description{
rename the file's name with the entered year and then read the file and store the data inside dat variable
It through warning when input is invalid
}
\references{
https://www.coursera.org/learn/r-packages/peer/25XGp/documenting-code
}
