% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{year}{input from user}
}
\value{
data table with file's data
}
\description{
Creates a summary table with the counts of car fatalities
}
\references{
https://www.coursera.org/learn/r-packages/peer/25XGp/documenting-code
}
