% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{integer or string specifying
year of the dataset ready to be read in.}
}
\value{
string or a vector of strings specifying the CSV file name.
}
\description{
rename a specific CSV file and print the the name after renaming
}
\examples{
\dontrun{make_filename <- function(2013)}

}
\references{
https://www.coursera.org/learn/r-packages/peer/25XGp/documenting-code
Make name for read-in dataset of specific year

Make name for the CSV file containing Americam public
yearly data regarding fatal injuries suffered in motor
vehicle traffic crashes.
}
